requirements.txt
Pillow==6.0.0
